requirements.txt
streamlit
pypdf
google.generativeai
chromadb==0.4.5
python-dotenv
pysqlite3-binary